Add movie details for Coach Carter
The movie is very accurate to the true story of Coach Ken Carter, who really did lock his undefeated 1999 team out of the gym for poor grades. The players on the team (like Timo Cruz) were fictional characters, but the coach's son, Damien, was real and did play for him. The film's famous "Our deepest fear" quote is often mistaken for a Nelson Mandela speech, but it was actually written by an author named Marianne Williamson.